% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_public_files}
\alias{my_public_files}
\title{Get a list of Public Files}
\usage{
my_public_files(limit = 100, token)
}
\arguments{
\item{limit}{how many files to return, by default 100}

\item{token}{an OAuth token for YDisk}
}
\value{
a dataframe
}
\description{
\code{my_public_files} uploads the files to the folder of YD
}
\details{
The function returns a dataframe with all the details for your public files
}
\examples{
my_public_files(limit = 1000, token = set_YD_oauth())
}
