% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_YD_oath}
\alias{set_YD_oath}
\title{Checks for YDisk token in the Global Environment}
\usage{
set_YD_oath()
}
\value{
a string or an error
}
\description{
\code{set_YD_oath} checks if OAuth token for Yandex Disk is set in the Global Environment
}
\details{
If the OAuth token for Yandex Disk API is set in Global Environment as YDisk,
or anything else there starts with OAuth y0_, the function returns it as the OAth token
}
\examples{
set_YD_oath()
}
