% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_YD_oauth}
\alias{set_YD_oauth}
\title{Checks for YDisk token in the Global Environment}
\usage{
set_YD_oauth()
}
\value{
a string or NULL with warning
}
\description{
\code{set_YD_oauth} tries to find Yandex Disk OAuth token in Global Environment
}
\details{
First, the function seeks the YDisk in Global Environment. If YDisk is absent in GE,
it checks if there is a variable starting with OAuth y0_ and takes it as YD OAuth.
}
\examples{
\dontrun{
set_YD_oauth()
}
}
