% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{upload_file_2YD}
\alias{upload_file_2YD}
\title{File uploader for Yandex Disk API}
\usage{
upload_file_2YD(
  path = "disk:/",
  yd_fname = NULL,
  disk_fname,
  overwrite = FALSE,
  YD_oath
)
}
\arguments{
\item{path}{path to folder where to upload, by default is disk:/ (should start with disk:/).}

\item{yd_fname}{the filename to use for the uploaded file at Yandex disk. If not set, it is the same as the disk_fname.}

\item{disk_fname}{the filename (to be uploaded) on your hard disk}

\item{overwrite}{if the file should be overwritten (by defaul = FALSE)}

\item{YD_oath}{an OAuth token for YDisk}
}
\description{
\code{upload_file_2YD} uploads the files to the folder of YD
}
\details{
The function uploads the file from hard disk to the specified folder (path) of YD.
}
\examples{
get_YD_files(path = "disk:/Загрузки/", disk_fname = "../../bookmarks.html",
             overwrite = TRUE, YD_oath = set_YD_oath())
}
