% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_YD_files}
\alias{get_YD_files}
\title{Files present at the provided YDisk path}
\usage{
get_YD_files(path = "disk:/", limit = 100, token)
}
\arguments{
\item{path}{a path at YDisk, it should start with disk:/}

\item{limit}{how many files to return, by default 25}

\item{token}{an OAuth token for YDisk}
}
\value{
a tibble
}
\description{
\code{get_YD_files} returns a list of files at the path provided
}
\details{
The path preset by default is disk:/ (root folder of your YDisk),
set own path starting with disk:/. If there are no folders at the indicated path,
the function return an empty tibble.
}
\examples{
\dontrun{
yd_oauth <- set_YD_oauth()
get_YD_files(token = set_YD_oauth())
get_YD_files(path = "disk:/Загрузки/", limit = 100, token = yd_oauth)
}
}
