% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_YD_folders}
\alias{get_YD_folders}
\title{Folders present at the provided YDisk path}
\usage{
get_YD_folders(path = "disk:/", limit = 100, token)
}
\arguments{
\item{path}{a path at YDisk, it should start with disk:/}

\item{limit}{how many folders to return, by default 100}

\item{token}{an OAuth token for YDisk}
}
\value{
a tibble
}
\description{
\code{get_YD_folders} returns a list of dirs at the path provided
}
\details{
The path preset by default is disk:/ (root folder of your YDisk),
set own path starting with disk:/. If there are no folders at the indicated path,
the function return an empty tibble.
}
\examples{
\dontrun{
yd_oauth <- set_YD_oauth()
get_YD_folders(token = yd_oauth)
get_YD_folders(path = "disk:/Загрузки/", limit = 5, token = set_YD_oauth())
}
}
